Rename: 장소 리스트 관련 컴포넌트 폴더 구조 변경
