+++ components/features/place-list/PlaceListItem.tsx
